restore install_times.json file in new installer
Recreates timer functionality from old installer in new installer.
Moves shared methods between the two installers into the new installer.
